requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.0.0
faker>=19.0.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0
langgraph>=0.1.0
pydantic>=2.0.0
requests>=2.28.0
flask>=2.3.0
langchain
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client